Debug: input data10

+++ src/omv/modules/py_tf.c
@@ -276,7 +276,7 @@ STATIC void py_tf_regression_input_callback(void *callback_data,
     if(params->input_channels == 1) {
         if (params->input_datatype == LIBTF_DATATYPE_INT8) {
 
-            int64_t *model_input_i64 = (int64_t *) model_input;
+            float *model_input_i64 = (float *) model_input;
             float input_scale = params->input_scale;
             int input_zero_point = params->input_zero_point;
